requirements.txt
asgiref==3.3.0
autopep8==1.5.4
dj-database-url==0.5.0
Django==2.2.10
djangorestframework==3.12.1
gunicorn==20.0.4
Pillow==8.0.0
pycodestyle==2.6.0
python-decouple==3.3
pytz==2020.1
sqlparse==0.4.1
toml==0.10.1
Unipath==1.1
whitenoise==5.2.0
